Make sure pid is set in array before issuing kill on it (avoid possibly killing self).

+++ src/slurmd/kill_tree.c
@@ -202,7 +202,8 @@ static void _kill_proclist(xpid_t *list, int sig)
 	while (list) {
 		verbose("Sending %d to %d", sig, list->pid);
 		/* Do not check errors. May already be dead */
-		kill(list->pid, sig);
+		if (list->pid > 0)
+			kill(list->pid, sig);
 		list = list->next;
 	}
 }
